Don't convert track art unless asked to
Previously in `getTracks`, we were encoding the artwork no matter what, even if we didn't include the artwork in the `fields` object param. This task is extremely expensive.

Here we move the task of encoding the image into section which checks the fields first. Now, we will only include and encode the image if no fields are passed or if the artwork field is explicitly passed.